Add PersonRepositoryTests, PersonServiceTests, OrdersServiceTests, update AuthTokenFilter, JwtUtils

+++ src/main/java/com/example/clothshop/controller/OrderController.java
@@ -1,18 +1,19 @@
 package com.example.clothshop.controller;
 
-import com.example.clothshop.dto.MapStructMapper;
-import com.example.clothshop.dto.OrderDTO;
-import com.example.clothshop.dto.ProductDTO;
-import com.example.clothshop.dto.OrderProductDTO;
+import com.example.clothshop.dto.*;
 import com.example.clothshop.entity.Orders;
 import com.example.clothshop.entity.Product;
+import com.example.clothshop.entity.Roles;
 import com.example.clothshop.service.OrderService;
 import com.example.clothshop.service.ProductService;
+import com.example.clothshop.service.UserDetailsImpl;
 import com.example.clothshop.util.exception.*;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,31 +31,50 @@ public class OrderController {
     private final MapStructMapper mapStructMapper;
 
     @GetMapping()
+    @PreAuthorize("hasRole('ADMIN')")
     public List<OrderDTO> getOrders() {
         return orderService.getOrders().stream().map(this::convertToOrderDTO).collect(Collectors.toList());
     }
 
     @GetMapping("/{id}")
-    public OrderDTO getOrder(@PathVariable("id") long id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public OrderDTO getOrder(@PathVariable("id") long id, Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, id);
         return convertToOrderDTO(orderService.getOrderById(id));
     }
 
     @GetMapping("/{id}/items")
-    public List<OrderProductDTO> getOrderItems(@PathVariable("id") long id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public List<OrderProductDTO> getOrderItems(@PathVariable("id") long id, Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, id);
         List<Product> products = orderService.getOrderProducts(id);
         return products.stream().map(mapStructMapper::productToOrderProductDTO).collect(Collectors.toList());
     }
 
     @GetMapping("/{oid}/items/{iid}")
-    public ProductDTO getItemOfOrder(@PathVariable("oid") long oid, @PathVariable("iid") long iid) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ProductDTO getItemOfOrder(@PathVariable("oid") long oid, @PathVariable("iid") long iid, Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, oid);
         Product product = orderService.getOrderItem(oid, iid);
         return mapStructMapper.productToProductDTO(product);
     }
 
     @PostMapping("/{id}/items")
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     public OrderProductDTO addItemToOrder(@PathVariable("id") long id,
                                           @RequestBody @Valid OrderProductDTO orderProductDTO,
-                                          BindingResult bindingResult) {
+                                          BindingResult bindingResult, Authentication authentication) {
+
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, id);
+
         Product product = productService.getProductById(orderProductDTO.getId());
         productService.checkForValidationErrors(bindingResult);
         orderService.addProductToOrder(id, product, orderProductDTO.getQuantity());
@@ -67,7 +87,12 @@ public class OrderController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable long id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public String delete(@PathVariable long id, Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, id);
+
         if (orderService.deleteOrder(id)) {
             return "Order with ID = " + id + " was deleted.";
         } else {
@@ -76,40 +101,76 @@ public class OrderController {
     }
 
     @DeleteMapping("/{oid}/items/{iid}")
-    public String deleteItemOfOrder(@PathVariable("oid") long oid, @PathVariable("iid") long iid) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public String deleteItemOfOrder(@PathVariable("oid") long oid, @PathVariable("iid") long iid,
+                                    Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, oid);
+
         orderService.deleteProductOfOrder(oid, iid);
         return "Product with ID = " + iid + " was deleted from order with ID = " + oid;
     }
 
     @GetMapping("/users/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public List<OrderDTO> getOrdersOfPerson(@PathVariable("id") long id) {
         List<Orders> ordersList = orderService.getOrdersOfUser(id);
         return ordersList.stream().map(mapStructMapper::orderToOrderDTO).collect(Collectors.toList());
     }
 
+    @GetMapping("/users")
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public List<PersonOrdersDTO> getOrdersOfUser(Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long id = userPrincipal.getId();
+        List<Orders> orders = orderService.getOrdersOfUser(id);
+        return orders.stream().map(mapStructMapper::orderToPersonOrdersDTO).collect(Collectors.toList());
+    }
 
-
+    /*
     @PostMapping("/users/{id}")
     public OrderDTO createOrderForPerson(@PathVariable("id") long id, @RequestBody @Valid OrderDTO orderDTO,
                                          BindingResult bindingResult) {
         orderService.checkForValidationErrors(bindingResult);
         return mapStructMapper.orderToOrderDTO(orderService.saveNewOrder(orderDTO, id));
     }
+    */
+    @PostMapping("/users/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
+    public OrderDTO createOrderForPerson(@PathVariable("id") long id) {
+        return mapStructMapper.orderToOrderDTO(orderService.saveNewOrder(id));
+    }
+
+    @PostMapping("/users")
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public OrderDTO createOrderForUser(Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long id = userPrincipal.getId();
+        return mapStructMapper.orderToOrderDTO(orderService.saveNewOrder(id));
+    }
 
     @PostMapping("/{id}/purchase")
-    public OrderDTO purchaseOrder(@PathVariable("id") long id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public OrderDTO purchaseOrder(@PathVariable("id") long id, Authentication authentication) {
+        UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+        Long personId = userPrincipal.getId();
+        orderService.checkAuthority(personId, id);
         return mapStructMapper.orderToOrderDTO(orderService.purchaseOrder(id));
     }
 
     @PostMapping("/{id}/cancel")
+    @PreAuthorize("hasRole('ADMIN')")
     public OrderDTO cancelOrder(@PathVariable("id") long id) {
         return mapStructMapper.orderToOrderDTO(orderService.cancelOrder(id));
     }
 
     @PostMapping("/{id}/ship")
+    @PreAuthorize("hasRole('ADMIN')")
     public OrderDTO shipOrder(@PathVariable("id") long id) {
         return mapStructMapper.orderToOrderDTO(orderService.shipOrder(id));
     }
+
 
     private OrderDTO convertToOrderDTO(Orders order) {
         return mapStructMapper.orderToOrderDTO(order);
@@ -141,7 +202,13 @@ public class OrderController {
 
     @ExceptionHandler
     private ResponseEntity<PersonErrorResponse> handlePersonNotFoundException(PersonNotFoundException exception) {
-        PersonErrorResponse response = new PersonErrorResponse("Person with this id wasn't found!");
+        PersonErrorResponse response = new PersonErrorResponse("User with this id wasn't found!");
         return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler
+    private ResponseEntity<PersonErrorResponse> handlePersonUnauthorizedException(PersonUnauthorizedException exception) {
+        PersonErrorResponse response = new PersonErrorResponse("User is not authorized to perform the requested operation.");
+        return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
     }
 }
